fix panic in typeck which was causing language service to go into a bad state (#1766)
As the title says, this fixes a panic that was causing the language
service to go into a bad state. We were assuming that all paths have
type schemes, but that's not the case as they can point to namespaces.
This should close #1762. Those failures in #1762 were arising from a
compilation update worker crashing (due to the aforementioned panic)
while holding a lock to the refcell.