Fix accumulating perf impact of `qsharp.run` with may shots in Python
The existing imlementation of `qsharp.run` would pass the entry expression for every shot executed, causing it to be recompiling and building up large internal state due to incrementing node ids and the like. This updates the patterns to have the interpreter track the last execution graph it compiled such that a new API in Rust can re-`eval` that last expression without incurring any additional compilation.