Move all samples testing to Rust unit tests (#1739)
This expands the unit tests in the samples_test folder to include
testing of all samples, including compilation of resource estimation
samples, compilation of dynamically discovered qsharp.json projects, and
compile and run of language samples. With these additions, the build
script no longer needs to compile the qsc cmdline tool and can skip the
cmdline based compilation of samples, so the build.py file is updated
here to remove those.