Add a qsharp.json context menu and palette command to add a project reference (#1771)
On invocation, a list of available projects to reference from the local
workspace and GitHub (the GitHub list to be populated as they become
available), or just a template for the user to complete.

On selection, the changes are made to the qsharp.json for the user to
review and save.

After invocation the drop-down appears similar to the below:

<img width="729" alt="image"
src="https://github.com/user-attachments/assets/3da3b381-0aab-4025-a327-f9867d5a38d4">