Noisy simulator performance optimization (#1764)
This PR introduces two main performance optimizations to the noisy
simulator:
- A `get_unchecked` access in the hot loop of the simulator to avoid
bound checks.
- A cache locality optimization: Because `nalgebra` stores its matrices
in column major form, it is very inefficient performing Matrix x Vector
multiplications, but fast performing Matrix.transpose() x Vector
multiplications. So, if we transpose all Kraus operators when they enter
the simulator, we can take advantage of this.

There are other minor perf improvements measured with `iai-callgrind`
that give us ~3% better performance each. Mostly operations that were
being done manually and now are being done with `nalgebra`.

Here is a screenshot of the total improvement in a 10-qubits and 10
operations circuit measured with `iai-callgrind`:

![image](https://github.com/user-attachments/assets/94a7de01-69d8-4e2e-97d5-64bc85976d12)