Fix bug for max_t_factories constraint (#1792)
A computation to determine the number of T factories was numerically
imprecise leading to results in which the constraint was not satisfied.

The computation has been fixed. Also the chemistry RE benchmark has been
updated for which this problem has been discovered in the first place.
Corresponding test cases for the chemistry RE benchmark are updated
accordingly.