Avoid errors in read-only GitHub sources (#1787)
This is an incremental improvement for viewing GitHub sources that
avoids treating them as open documents since they are read-only buffers.
In the future, we can do the work to associate them with the existing
compilation taht contains those sources so that things like hover and
goto-def work, but for now those code paths are user-code specific and
need some reworking to handle navigation in a compile unit for a
dependency.