Move `Unstable` into its own library (#1802)
This is branched off of #1801.

This PR makes a few changes around our library code:
- in `library/src/tests/resources`:
- turned into a Q# project so the language service works while editing
it
- the `Unstable` library:
	- modernized opens into imports with the new std API
	- removed explicit namespaces
- language service:
- Remove special casing around ignoring completions that are from
`Unstable`. Now, if the user is explicitly adding the unstable
dependency, they likely want completions.

After this PR goes in, we will no longer make any updates
`Microsoft.Quantum.Unstable`, and will instead use the
`library/unstable` library. This will allow our users to pin to a
specific version of `Unstable` via the external dependencies system.

The pipeline won't run for this as it isn't based off of `main`, but
`./build.py` is currently passing as of 4138fad.

---------

Co-authored-by: Scott Carda <Scott.Carda@microsoft.com>