Fix `noisy_simulator` panic when initializing with non-square Kraus matrices (#1826)
When initializing an `Operation`, the first thing we do is transposing
the matrices, `nalgebra` panics internally when transposing non-square
matrices. This PR moves the transposition operation after the dimensions
check to gracefully handle that case and adds a unit test to check for
this in the future.

Notes:
 1. This bug was introduced in the performance improvement PR.
2. It doesn't affect well-formed programs; it just affects programs
using ill-formed Kraus operators.