Fix duplicate imports being generated when using completions (#1820)
Closes #1777
Closes #1775 

This PR updates completions in the following ways:
- Organizes `completions.rs` a bit -- it had some very, very nested code
that was hard to make sense of.
- Doesn't insert an import if a symbol has already been imported via a
glob import or open
- Doesn't insert an import if a symbol has already been imported by name
exactly
- Recognizes if the parent namespace has already been imported via an
alias
- Removes the `ignore_internal_callable` test, which was testing
`internal` as if it would make things private to a single namespace,
when actually `internal` means private to a package.


Recognizing a preexisting glob import:
 


https://github.com/user-attachments/assets/2732da7a-8211-4bb3-927e-1af234b81fb8


Recognizing a preexisting alias:


https://github.com/user-attachments/assets/ca5e99f4-294a-4eed-8b91-d9133a08d124