Include completions for items from the same file; Clean up completions generation for callables (#1863)
This PR closes #1862.

I had previously touched this function in #1820. I did my best to port
it and maintain existing functionality, but there was one conditional
that I messed up, and it was around callables originating from the same
file.

This PR fixes the bug that caused. But also, this PR tries to simplify
`callable_decl_to_completion_item` so this is less likely to happen in
the future.