Add edit/run for kata examples - issue 591 (#1829)
Add the edit/run for kata examples.
This PR closes https://github.com/microsoft/qsharp/issues/591

Monaco editor model should have a unique URI (with few exceptions). So,
either kata lesson ID or kata exercise ID is provided to the model. The
Editor.tsx implementation is replacing "kataExercise" parameter with
"kataSection" that contains both lesson.id and exercise.id.
The function parameter "key" is not in use and was removed.