generate auto imports for core callables (#1861)
Closes #1860

Currently, we have a separate code path for generating completions for
core callables. This code path just hard-codes the callables to
not include any auto-generated imports. This may have been necessary at
some point, but it no longer is and it is preventing us from generating
the correct auto-imports for `QIR.Runtime`. We now have a function that
generates completions from a package and is aware of if imports are
needed or not, so I just went ahead and used that instead of special
casing `Core`.

This PR:
- Uses the standard completion codepath to generate completions for
`Core`
- Adds the ability to unit test completions with the real stdlib, not
just the fake stdlib
- Adds tests to ensure auto-imports are generated for items that are not
already in scope (e.g. `QIR.Runtime.*`) and are _not_ generated for
items already
  in scope (e.g. `Length`).