Implements serialization for physical resource estimation (#1892)
This adds serialization for the raw physical resource estimation result.
This improves exporting the result when working with resource estimator
through the extensibility API.

This does not affect the specialized resource estimation result for the
system implementation used in the RE integration into the VS Code plugin
and Python library.