Generic code with code distance and threshold (#1896)
Provides a generic version for an error correction code that has a code
distance, threshold, and exponential error suppression. The system
version (with surface code and Floquet code as defaults, and
customizable via formula strings) is now implemented based on the
generic version. The generic version can also be used in other projects
via extensibility API.