Basic interop with Qiskit (#1899)
This PR adds the ability to resource estimate, run, and generate QIR
from Qiskit `QuantumCircuit`s.

There are three areas which we can break up the review into:
- compiler/qsc_qasm3
- Python user facing API and notebooks. Example usage can be seen in:
  - samples/python_interop/qiskit/interop.ipynb
  - samples/estimation/estimation-qiskit.ipynb
- /compiler changes (excluding qsc_qasm3)

The QASM parser doesn't have usable semantic validation, so we have to
do a lot of our own validation as we compile the QASM generated from the
`QuantumCircuit`s.

Closes #1074

---------

Co-authored-by: orpuente-MS <156957451+orpuente-MS@users.noreply.github.com>
Co-authored-by: Mine Starks <16928427+minestarks@users.noreply.github.com>
Co-authored-by: Alex Hansen <alex@alex-hansen.com>
Co-authored-by: Stefan J. Wernli <swernli@microsoft.com>