Basic samples for RE API (#1915)
This adds to basic examples that illustrate how to use the RE API for
extensibility. These samples only use existing models that are used in
the RE integration with the QDK. Future samples will build on top of
these and illustrate how to change models for qubits, error correction
codes, factories, and post-layout logical overheads.