Serialize logical post-layout overhead in resource estimation result (#1914)
This change allows to serialize the logical post-layout overhead in the
resource estimation result provided by the resource estimation API. Note
that this was already possible in the system API that has a custom
result for the VS Code and Python integration.

The change also simplified the type of the resource estimation result,
which now is not any longer generic over the post-layout overhead trait.