Migrate the standard library to the project system (#1912)
This PR migrates the standard library into a Q# project structure.

Previously, we provided the modern standard library API via re-export,
while the "true" stdlib was defined in the old explicit-namespace style.
Now, that is inverted: the stdlib's code is housed in a project with
implicit namespaces, and the legacy (`Microsoft.Quantum.____`) API is
provided via re-exports.

Why do this? Well, for one, the standard library should look like a
regular library and use the latest features we have.
Secondarily, auto-imports and other language dev tools now refer to
these items by their `Std.__` name, not their `Microsoft.Quantum.__`
name. See the video below for an example of this in completions:


https://github.com/user-attachments/assets/fa96e5b7-0cca-4a44-9224-7742d50f8370

Go-to def rendering the new paths:


https://github.com/user-attachments/assets/5fea33ec-9478-4c16-9e78-b8f8b1420ed0



Note that `Unstable.*` has not been migrated to the modern API, as we
want to explore stabilizing state preparation (#1910)

Core is also not included, but for different reasons: #1911. Coming in a
Follow Up PR™️ (`alex/1911` if you want to preview the work there)