Migrate core library to the new Stdlib API/projects system (#1919)
This is a follow-up to #1912, and includes an associated RCA bug fix
(#1918 ) that was not exposed until I tried to re-export something from
`Core`.

This also makes a subtle change to the compiler where the prelude is no
longer created by default in the namespace tree, and is rather created
when we compile `core.qs`. This was needed because we no longer allow
"clobbering namespaces", that is, re-opening a namespace from an
_external_ package.

---------

Co-authored-by: César Zaragoza Cortés <cesarzc@microsoft.com>