Optimize RIR reindexing, QIR qubit use (#1938)
This change updates the behavior of the RIR qubit reindexing to assume
single block only and then take additional steps to optimize qubit usage
with that block. This is consistent with the assumption that qubit reuse
is required for adaptive programs rather than a separately configurable
capability.

In addition, updates the qasm3 compiling to take advantage of this
optimization by having explicitly call `__quantum__qis__m__body`, and
updates the corresponding tests.