Update circuits widget sizing behavior (#1921)
This change fixes a few quirks of the `Circuit` widget, specifically:

- if the width of the widget is greater than the width of the cell
output area, scroll bars are added, and
- setting a custom zoom level by changing the value in the zoom control
disables the resize handling so that VS Code doesn't incorrectly trigger
a recalculation of the default zoom and undo the manual change.

Confirmed with manual testing that window resizing in VS Code circuit
display is still hadled properly.

---------

Co-authored-by: Mine Starks <16928427+minestarks@users.noreply.github.com>