Control how physical qubits are computed in factories (#1940)
This adds the possibility to change the behavior, how physical qubits
are computed in a round based factory. The default behavior is to take
the max among all distillation rounds, but one can set to calculate the
sum over all units instead. This change is accessible via the job
params.