Use all github dependencies in published libraries (#1956)
I was just importing FixedPoint and noticed it doesn't work due to a
local path dependency. This fixes that, but also updates the refs for
other libraries to latest.

Aside: I can't believe this got checked in with an absolute path on my
PC. That's a big L on my part.