Fix to partial evaluation generating branch instructions on constant conditions (#1963)
RCA can categorize an expression as dynamic even if it is in fact purely
classical. This can happen in cases where a data structure such an
array, tuple or UDT contains a mix of static and dynamic values. In such
instances, RCA identifies all the contents of the data structure as
dynamic even if some values are static. This causes partial evaluation
to generate branch instructions on constant conditions under certain
circumstances. This change fixes this issue.

---------

Co-authored-by: Stefan J. Wernli <swernli@microsoft.com>