Add custom measurement operations to Q# (#1967)
This PR adds the ability to declare custom measurements in Q#. Here is
an example:
```qsharp
@Measurement()
operation __quantum__qis__mx__body(target : Qubit) : Result {
  body intrinsic;
}
```

The generated QIR will be,
```llvm-ir
declare void @__quantum__qis__mx__body(%Qubit*, %Result*) #1
```

# Design Notes

1. I choose the `@Measurement()` attribute over a new `measurement`
keyword to avoid putting extra syntax into the language.
2. A custom measurement must be a `body intrinsic` or a
`@SimulatableIntrinsic()`.
3. A measurement must only take `Qubits` as inputs and have `Results` as
outputs.

# Notes to reviewers

This PR consists of the following steps:
1. Adding a `@Measurement()` attribute to the language and piping it
down to FIR.
3. Make the generated QIR have the `#1` attribute indicating an
irreversible operation.
 4. Treating a measurement as a dynamic operation in RCA.
5. Handling custom measurements correctly during partial evaluation:
moving the output Results to result registers in the input.
 6. Handling undesired semantics as errors using an HIR pass.
 7. Adding the `@Measurement()` attribute to the completions list.
 8. Unit tests in codegen and in the HIR pass.
 
 # Next steps
 
After this PR is merged we can add the `@Measurement()` attribute to the
measurement operations in the standard library and remove the hard coded
logic we have for them in the compiler.

---------

Co-authored-by: Stefan J. Wernli <swernli@microsoft.com>
Co-authored-by: César Zaragoza Cortés <cesarzc@microsoft.com>