`DumpMachine` output in Python and console should be empty with no qubits allocated
This change updates the display of quantum state when no qubits are allocated to show a message rather than the same output as one qubit in the zero state. This matches the updated debugger behavior. Note that I've tried to ensure this is NOT a breaking change for consumers of the WASM/npm package by extending the data type rather than changing the behavior or contents of existing fields, and verifyied this by leaving the dev playground unchanged and confirming it preserved the old behavior there (playground will be updated in a follow up PR and is tracked by a separate issue).

Fixes #1961