Skip checking stack frames in evaluator expect tests
It used to be that we avoided noise in the evaluator tests during libraries edits by avoiding having spans from the stdlib show up in there, but newer tests include calls to `Std.Intrinsic.I` that need spans updated every time something alphabetically above that operation changes in the stdlib. This change updates the expect call on failure to only verify the error and not the stack frames, as call stacks are already verified in other tests and don't need to be checked here.