Fix RCA panic (#2017)
Fixes #1925

# Notes for reviewers
The issue was that a helper function analyzing binary operations was
assuming that operands were element values (non-arrays). I just updated
it to also consider arrays.

The logic I added to handle binary ops in arrays is marking the output
of a binary operation as `Dynamic` if either:
 1. The elements of the array are `Dynamic`.
 2. The size of the array is `Dynamic`.

`Dynamic` in this context means that it cannot be known at compile time.