Add class constraints for built-in classes (#2007)
This PR exposes some relevant built-in classes via class constraints.

# Changes
- Add `TypeParameter`, `ClassConstraint`, and `ConstraintParameter` AST
nodes to represent the notion of a generic type that might have class
constraints, and constraints that may have parameters (e.g. `Exp[Int]`).
  -  update visitors accordingly
- `Ty` can now contain arbitrary constraints
- Add notion of primitive classes to the AST
- Update constraint checking algorithm to check type parameters for
primitive constraints at callable decl time
- Update constraint generation algorithm to add additional
user-specified constraints to arguments that are passed in to callables
(`fn constrained_ty`)
- Add completions for primitive classes to the language service
- Adds a sample describing class constraints

tiny changes:
- Add docstrings discretionally
- Remove some unnecessary derives
- Tried to clarify and unify jargon around classes, constraints,
parameters, and arguments in general
- Some expect_tests were updated in minor ways as their underlying types
changed (`Ty` and `TyParam`, for example)
- Improve an error message for `MissingTy` (it mentioned something about
global types not being inferred when 99% of the time it does not apply
to a global type, it applies to a callable parameter with a missing
type)