Track qubit live-ness during simulation (#2020)
This change adds some basic qubit live-ness tracking to the simulator,
achieved by changing the `Value::Qubit` variant to a weak pointer to the
qubit and having the execution environment own and track the `Rc` that
owns the qubit. This provides a few benefits:

- Fixes a panic when a qubit is passed to the simulator after it has
been released.
- Adds graceful error handling to treat use-after-release as a runtime
error.
- Adds a specific error for "double release" of a qubit if it is being
manually managed.
- Updates the display of the qubit type to show the qubit ID when it is
live and `Qubit<released>` when it is not, enabling easier debugging of
qubit lifetimes.

Of note, local perf testing showing no impact on performance (thanks,
Rust zero-cost abstractions!), though it stands to reason this change
could theoretically have some impact on programs that use many thousands
of qubits.