Moved formerly unstable to Std library (#2022)
We had Arithmetic, Arithmetic Utils (internal), Table Lookup and State
Preparation in the unstable library. Now they are moved to the Std
library.

Taking Arithmetic as an example, we now place Q# code in the
Std.Arithmetic and re-export it in Microsoft.Quantum.Unstable.Arithmetic
for compatibility. Separate _unstable_ library is removed.

We probably should look at using AND gate in subsequent PR.

Co-authored-by: Dmitry Vasilevsky <dmitryv@microsoft.com>