Fix Qiskit interop with v1.3 (#2050)
Qiskit v1.3 changed the behavior of `num_qubits` on the `Target` class
in way that broke our `QirTarget` usage, since we expect to default to
`None` but the new logic ignores the `num_qubits` parameter to
`__init__` and always defaults to zero qubits, failing later calls to
transpilation. By shadowing the parent class property in `QirTarget` we
can continue returning `None` and trigger the right behavior in
transpilation while preserving a setter for use with v1.2.

Note: I tested locally with both v1.2.4 and v1.3.0 integration tests to
confirm the behavior is as expected.

Fixes #2047