Switch QIR Target checking from allow list to deny list (#2058)
This change updates the logic we use to decide if a target supports QIR
submission in the extension. Previously it was an allow list which
required any new targets or providers to be explicitly added. Instead,
this uses a deny list style so that new providers and targets are
allowed to submit either base or adaptive without requiring updates to
the QDK. Then we can coordinate with new partners on what targets should
be lit up for which purposes.