Syntax highlighting for functions, variables and numbers (#2088)
Supports syntax highlighting for:

- function definitions
- function calls
- numbers
- variables

See issue https://github.com/microsoft/qsharp/issues/1950.

The change for function calls also applies to operation definitions and
to attributes (see bigger example below).
The change for variables also applies to attributes and to imports.

I might be nice to make attributes and imports get their own designated
rules. I can do them as part of this PR, but I think it might be best to
do as a followup issue - because I believe this PR is actually already
an improvement for them.

Simple example for the change (from the original issue description):


![image](https://github.com/user-attachments/assets/cab2630b-0e28-4d48-a98e-7e5bc83e5c7c)

And a bigger example:


![image](https://github.com/user-attachments/assets/7d0c059a-8481-4159-ba64-d8539f1540c0)