Fix `Exp` on qubit arrays larger than 2 with single `PauliI` (#2086)
This fixes a bug in the decomposition for `Exp` that would cause a
runtime failure if an array of more than two qubits was used and a
single `PauliI` was passed in the Pauli array. The decomposition ignores
`PauliI` by calling the `RemovePauliI` helper function, but then
incorrectly uses the original arrays when rather than the filtered
arrays in the rest of the operation. This adds a test to confirm the fix
(without the fix the test triggers a runtime failure in `MapPauli`).