Adaptive profile floating-point computations extension (#2078)
Adds dynamic float support implementing the floating-point computations
for the AdaptiveRIF profile. This also implements `fcmp` missing from
the spec which was [filed as a spec
bug](https://github.com/qir-alliance/qir-spec/issues/41).

The `fcmp` instructions chosen were the ordered set. The spec says
operations like dividing by `0` are undefined behavior and diving by `0`
usually gives a `NAN` value. For any `fcmp` calls we need to choose
ordered or unordered. I'm not sure that it matters which we decide.

This PR does not change the default profiles for the defined hardware
targets.