Added simple VQE sample (#2073)
This PR adds simple self-contained example of a Variational Quantum
Eigensolver (VQE).

This example includes:

1. Simple classical optimization to find minimum of a multi-variable
function in order to find the approximation to the minimum eigenvalue of
a Hamiltonian.
2. Finding Hamiltonian expectation value as a weighted sum of terms.
3. Finding one term expectation value by performing multiple shots of
preparing ansatz state and measuring it.
4. Ansatz state preparation similar to the circuit in the published
paper "Ground-state energy estimation of the water molecule on a trapped
ion quantum computer" (https://arxiv.org/abs/1902.10171)

To keep this sample simple and self-contained Hamiltonian is artificial
(doesn't correspond to any actual chemistry problem). Also, this sample
doesn't include any application of VQE, for example, finding a geometry
of a molecule.

---------

Co-authored-by: Dmitry Vasilevsky <dmitryv@microsoft.com>