Fix global phase for controlled-T, R1 (#2112)
The decomposition for controlled-`T` in the stdlib uses `Rz` gates that
are correct up to a global phase, which is fine for hardware but
distracting in simulation. Correcting for that global phase in `T` fixes
the global phase in `S`, which in turn allows removing a patch that made
`R1` use a global phase consistent with `S`.

Fixes #2106