Fix language service panic when file isn't listed in the `files` field of `qsharp.json` (#2109)
Fixes #2090

Notably, this also changes how we treat the `files` field in
`qsharp.json` for local projects their local dependencies.

Previously, if the `qsharp.json` explicitly listed any `files` , that
list would take priority and we would skip crawling the directory.

With this change: we always crawl the `src/` directory and include all
found `*.qs` files in the project. But we raise an explicit error to the
user if any of those files are missing from the `files` list.