Update test callable discovery to work within the language service (#2095)
This PR rewrites #2059 to trigger test discovery from within the
language server, as opposed to externally in an entirely separate
context. Please read the description in that PR for more detail.

It also adds tests for LS state in both the Rust-based LS tests and the
JS-based `basics.js` npm API tests.

---------

Co-authored-by: Mine Starks <16928427+minestarks@users.noreply.github.com>