`Relabel` in adaptive conditional block should be disallowed (#2155)
This adds a check at codegen time if a qubit relabel is occurring in an
adaptive conditional. This will ensure we don't generate QIR with
incorrect qubit IDs. The change ties the error to the dynamic qubit
capability because if/when we support dynamic qubits we'll need to
update relabel to become an intrinsic that is supported by backends to
do a runtime relabel of qubit identifiers.

Fixes #2152