Chemistry library, SPSA sample, and notebook (#2105)
Converting the old chemistry library and Variational Quantum Algorithms
sample to modern QDK. As part of this, I tried where possible to update
to our newer Q# programming patterns. I've tried to make sure we export
on the reasonable part of the APIs as public facing, but if there is
something that is not exported and should be or something that is
exported and shouldn't be we can adjust.

For the original library code, see
https://github.com/microsoft/QuantumLibraries/tree/main/Chemistry/src/Runtime.
For the original sample notebook, see
https://github.com/microsoft/Quantum/blob/main/samples/algorithms/variational-algorithms/Variational%20Quantum%20Algorithms.ipynb

---------

Co-authored-by: DmitryVasilevsky <60718360+DmitryVasilevsky@users.noreply.github.com>
Co-authored-by: Dmitry Vasilevsky <dmitryv@microsoft.com>