Update registry for chemistry (#2161)
Also changing the 'stable' reference to point to a tag. (We should
probably auto-bump 'stable' to be the latest release tag for each
release - I'll add a task for that.

Chemistry wasn't in v1.13, so just pointing that to main for now, and
will add a 'stable' tag for next release.