LS: Hide errors from github sources only when they're not associated with a real project (#2139)
Small fix to #1787 while staying true to the intention of that change.

We want to _hide_ errors from GitHub sources that are open in the
editor, but aren't associated with a project. They're essentially
orphans. For example, you may have closed the original project that
referenced those GitHub sources, but left the GitHub source open in the
editor.

But we want to _show_ errors from GitHub sources if they're part of a
normal project. Errors from dependencies are helpful.