Support code lenses on more callables
This change adds support for code lenses on any callable that qualifies for them instead of just the entry point. In practice, this means any function or operation that doesn't take arguments can be used as the starting place for a run, debug, estimate, histogram, or (in the case of operations) circuit. This is paticularly helpful for unit tests identifed by the `@Test()` attribute, as it lets them be run or debug outside of the test infra with a single click.