Fixes possible infinite loop in RE (#2128)
In the rare case that a factory runs longer than the algorithm, we
adjust the number of cycles. The heuristic we use is to take the maximum
number of cycles for the current code parameter + 1. However, for some
code definitions this may end in an infite loop, which we break by
ensuring that the number of cycles is always increasing.