Fix RCA panic on lambda with explicit return (#2194)
This fixes a quirk leftover from the fixes in #649 and #793. Those
introduced and then updated a conditional that avoided unifying certain
types during type checking, but it turns out it was always safe to
unconditionally unify the types once #649 introduced the explicit return
type tracking. With this fix, the lifted lambda callables with explicit
returns have the correct output type instead of always being `Unit` and
RCA is able to use that type information as expected, avoiding the
panic.

Fixes #2186