Multi-target gate circuit art improvement (#2185)
This PR has a proposed fix for issue
https://github.com/microsoft/qsharp/issues/2184, but it mainly includes
refactoring of the circuit-art code based on comments from PR
https://github.com/microsoft/qsharp/pull/2126.

The only real functionality change in this code is that multi-target
gates get extra lines that behave in the same way as measurement lines,
in order to better convey the qubits involved.

For example, a circuit like this one:
```
use q0 = Qubit();
use q1 = Qubit();
Rxx(1.0, q0, q1);

use q2 = Qubit();
use q3 = Qubit();
Rxx(1.0, q2, q3);
```
instead of rendering as
```
q_0    ─ rxx(1.0000) ─
q_1    ─ rxx(1.0000) ─
q_2    ─ rxx(1.0000) ─
q_3    ─ rxx(1.0000) ─
```
will instead render as 
```
q_0    ─ rxx(1.0000) ─
              ┆
q_1    ─ rxx(1.0000) ─
q_2    ─ rxx(1.0000) ─
              ┆
q_3    ─ rxx(1.0000) ─
```

This is very similar to what we get if we measure the first qubit of
each of these two-qubit gate, i.e.
```
q_0    ─ rxx(1.0000) ─── M ──
              ┆          ╘═══
q_1    ─ rxx(1.0000) ────────
q_2    ─ rxx(1.0000) ─── M ──
              ┆          ╘═══
q_3    ─ rxx(1.0000) ────────
```

On top of this change in rendering, the first few commits in this PR are
actually just code refactoring. As such, it's probably best to review
this PR commit by commit, for two reason - a. it's going to be much
easier to follow, and b. any of these commits can be reverted if you
think they make the code more complicated instead of making it easier to
read.

The commits are as follows:
1.
[a400655](https://github.com/microsoft/qsharp/pull/2185/commits/a4006557b4b0fcd0bacd14c5c1142eeecefbaa76)
A rework of PR https://github.com/microsoft/qsharp/pull/2126, which, to
be honest, is how I think I should have implemented it in the first
place.
2.
[c141799](https://github.com/microsoft/qsharp/pull/2185/commits/c14179907ed45ebd02e07199db85a401edaf61b1)
A cleanup based on [this
comment](https://github.com/microsoft/qsharp/pull/2126#pullrequestreview-2579485021).
It makes the code significantly better, in my opinion.
3.
[4012ee8](https://github.com/microsoft/qsharp/pull/2185/commits/4012ee8378f0080cf79b2220dead041ccc964edf)
A small simplification to have the ColumnRenderer have a `new` method
which is in charge of making sure that the column width is an odd
number, and also a `Default` trait implementation for the case where we
don't need a specific width.
4.
[cfa7382](https://github.com/microsoft/qsharp/pull/2185/commits/cfa7382630b83ee0aefd64ee414ca6b5b600f2d2)
This is the actual enhancement - it makes sure there is a classical line
(same as what you get from a measurement) if there are two consecutive
qubits for the same operation. If you think there is a better solution
to make the rendering clearer, let me know. But of the ones I could
think of, I found this one to be the least disruptive one (to both the
existing code, and the existing circuits).